refactor: Clean up header
