+++ src/shared/components/Header/Header.js
@@ -1,6 +1,6 @@
-import React, { Fragment, useEffect, useState } from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import { useHistory } from 'react-router-dom'
+import React, { useState } from 'react'
+import { useSelector } from 'react-redux'
+import { selectHeaderInfo } from 'features/gameSlice'
 import { HeaderMenu } from '../HeaderMenu'
 import {
   HeaderWrapper,
@@ -10,57 +10,28 @@ import {
   IconBell,
   IconUser
 } from './styled'
-
-import { selectHeaderInfo, playerUpdated, logoutGame } from 'features/gameSlice'
-import Api from 'services'
+import utils from 'utils'
 
 export const Header = () => {
-  const dispatch = useDispatch()
-  const { username, headerTitle, isLoggedIn } = useSelector(selectHeaderInfo)
+  const { username, headerTitle } = useSelector(selectHeaderInfo)
   const [openMenu, setOpenMenu] = useState()
-  const history = useHistory()
-
-  const firstChar = () => {
-    if (typeof username !== 'string') return ''
-    return username.charAt(0).toUpperCase()
-  }
-
-  useEffect(() => {
-    Api.getMe().then(data => {
-      dispatch(playerUpdated(data.username))
-    }).catch(() => {
-      dispatch(logoutGame())
-    })
-  }, [dispatch])
-
-  const quitRoom = () => {
-    (() => {
-      history.push(isLoggedIn ? '/rooms' : '/')
-    })()
-  }
 
   return (
-    <HeaderWrapper show={isLoggedIn}>
+    <HeaderWrapper>
       <NavBar>
-        {!isLoggedIn && <div />}
-        {isLoggedIn &&
-          <Fragment key={0}>
-            <Text onClick={quitRoom}>{headerTitle}</Text>
-          </Fragment>}
+        <Text>{headerTitle}</Text>
       </NavBar>
 
-      {isLoggedIn &&
-        <div>
-          <ProfileContainer>
-            <IconBell type='bell' size={0.21} />
-            <IconUser
-              onMouseEnter={() => setOpenMenu(true)}
-              onMouseLeave={() => setOpenMenu(false)}
-            >{firstChar()}
-            </IconUser>
-          </ProfileContainer>
-          <HeaderMenu open={openMenu} />
-        </div>}
+      <ProfileContainer>
+        <IconBell type='bell' size={0.21} />
+        <IconUser
+          onMouseEnter={() => setOpenMenu(true)}
+          onMouseLeave={() => setOpenMenu(false)}
+        >
+          {utils.firstChar(username)}
+        </IconUser>
+      </ProfileContainer>
+      <HeaderMenu open={openMenu} />
     </HeaderWrapper>
   )
 }
