+++ src/shared/components/Header/styled.js
@@ -1,5 +1,4 @@
-import styled, { css } from 'styled-components/macro'
-import { imageCDN } from 'mixins'
+import styled from 'styled-components/macro'
 import { sizes, font, mixin } from 'shared/utils/styles'
 import { Icon } from '../Icon'
 
@@ -7,33 +6,6 @@ export const HeaderWrapper = styled.div`
   width: 100%;
   height: ${sizes.bannerHeight}rem;
   background: #FFFFFF;
-
-  ${props => !props.show && css`
-    background: #FAFBFD;
-    display: none;
-  `}
-`
-
-export const Arrow = styled.div`
-  ${props => props.sidebar
-    ? imageCDN('icon-arrow-left.png', 0.38, 0.38)
-    : imageCDN('icon-arrow-right.png', 0.38, 0.38)}
-  transition: background-image 0.3s;
-  margin-right: 0.05rem;
-`
-
-export const Logo = styled.div`
-  ${imageCDN('game-logo.png', 0.33, 0.33)}
-  margin-top: 0.02rem;
-  margin-right: 0.07rem;
-  display: inline-block;
-`
-
-export const MainText = styled.div`
-  font-family: 'Orbitron', sans-serif;
-  font-size: 0.3rem;
-  font-weight: bold;
-  display: inline-block;
 `
 
 export const Text = styled.div`
@@ -52,7 +24,6 @@ export const NavBar = styled.div`
   position: absolute;
   top: 0.05rem;
   margin-left: 0.3rem;
-  cursor: pointer;
   display: flex;
 `
 
@@ -63,25 +34,6 @@ export const ProfileContainer = styled.div`
   justify-content: center;
   align-items: center;
   flex-direction: row;
-
-  .info {
-    margin-top: 0.09rem;
-    width: 1.4rem;
-    text-align: left;
-    .name {
-      color: #000;
-      font-weight: bold;
-      font-size: 0.17rem;
-      margin-left: 0.08rem;
-      text-align: left;
-    }
-  
-    .points {
-      color: #424242;
-      font-size: 0.14rem;
-      margin-left: 0.08rem;
-    }
-  }
 `
 
 export const IconUser = styled.div`
