+++ src/shared/components/HeaderMenu/HeaderMenu.js
@@ -10,6 +10,7 @@ import {
 } from './styled'
 import { logoutGame } from 'features/gameSlice'
 import toast from 'shared/utils/toast'
+import utils from 'utils'
 
 export const HeaderMenu = ({ open }) => {
   const dispatch = useDispatch()
@@ -19,6 +20,7 @@ export const HeaderMenu = ({ open }) => {
   const logout = async () => {
     try {
       dispatch(logoutGame())
+      await utils.delay(50)
       history.push('/')
     } catch (err) {
       toast.error(err.response?.data?.msg || 'error-message')
